New roles, frontend !!! ADD ROLES MANUALLY  TO DB!!!

+++ source/backend/src/main/java/cz/osu/pizzakaktus/repositories/UserRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
  * Created by e-myslivost on 26.10.2016.
  */
 @Repository
-public interface UserRepository extends CrudRepository<UserDb, Long> {
-    List<UserDb> findById(Long Id);
+public interface UserRepository extends CrudRepository<UserDb, Integer> {
+    List<UserDb> findById(Integer Id);
     List<UserDb> findByLogin(String login);
 }
